explicit cast from double to float, to avoid compiler warning

+++ core/vgl/vgl_homg.cxx
@@ -8,8 +8,8 @@
 # include <math.h> // dont_vxl_filter: no HUGE_VAL in <cmath>
 #endif
 
-float vgl_homg<float>::infinity = HUGE_VAL;
-float vgl_homg<float>::infinitesimal_tol = 1e-12;
+float vgl_homg<float>::infinity = (float)HUGE_VAL;
+float vgl_homg<float>::infinitesimal_tol = 1e-12f;
 
 double vgl_homg<double>::infinity = HUGE_VAL;
 double vgl_homg<double>::infinitesimal_tol = 1e-12;
